Applied Crypto++: Pseudo Random Number Generators By Jeffrey Walton, 10 Apr 2008
Random numbers are a primitive element not only to cryptographers, but Computer Scientist and Programmers in general. What is desired is a method which produces a pseudo random stream of numbers fast which are cryptographically secure. However, the goals are diametrically opposed - pseudo random sequences can be produced quickly, or can be produced strongly; but usually not quickly with properties which resist Cryptanalysis.

http://www.codeproject.com/Articles/16982/Applied-Crypto-Pseudo-Random-Number-Generators